requirements.txt
# ARGO Phase 2 Requirements

# Core dependencies
redis==5.0.1
python-dotenv==1.0.0
pyyaml==6.0.1

# Async support
asyncio==3.4.3
aioredis==2.0.1

# LLM APIs (install as needed)
openai==1.6.1
anthropic==0.8.1
google-generativeai==0.3.2

# Logging and monitoring
python-json-logger==2.0.7
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Development tools
black==23.12.1
flake8==7.0.0
mypy==1.7.1